dbushook.cpp, dbuspdo.cpp, dbusprxy.cpp, iomgr.cpp: removed redudant dbus_connection_close.  Now we keep just one close in the CDBusBusPort's ReleaseDBus, to make things simpile, that is, the dbus watch remove can only happen when the CDBusBusPort's is stopping, otherwise it could happen in one of the PDO's of the CDBusBusPort.
irp.cpp: fixed a bug in irp timeout handler when the CancelIrp should be called
instead of CompleteIrp, which has an operation to remove the irp reference in
the port's maps, while CompleteIrp does not have this operation. Use of the
CompleteIrp can cause IRP to leak.

iomgr.cpp: fixed a bug in CancelIrp to return the specified return value, rather
than the return code of CPort::CancelIrp.

rpcif.cpp: fixed a bug in CRpcInterfaceBase::AddAndRun to return STATUS_PENDING
if the paralell task group cannot be run immediately, which was not set and
    caused the endless wait.

pnpmgr.cpp: fixed a bug in CPnpMgrQueryStopCompletion, that IRP to reuse is not
cleared completely. Now the irp is retired and a new Irp is used for the new
request.

tcpport.cpp: fixed a bug in CRpcStreamSock's GetStreamFromIrp, where the stream
id should be retrieved from the propStreamId, but not property 0, which caused
many irp leaks.

msgmatch.h: fixed a error, that dbus_message_iter_get_basic is fed with a
nullptr.

dbusprxy.cpp: fixed a bug that the dbus boolean width is 4 bytes and was passed
a c++ bool variable to hold the dbus boolean variable.

tasklets.cpp: fixed a concurrent bug in CSyncCallback, when the samaphore is
posted before the error code is set. in some busy condition, the semaphore
waiter could get a stale error code rather than the intend one.

tasklets.h: changed the CTasklet::m_iRet to an atomic variable, so that the
GetError can be called safely from different threads. There could be chances the
SetError and GetError runs at the same time from different processors.

prxyfdo.cpp, prxyfdo.h: added SetConnected, IsConnected to CDBusProxyFdo and
m_matchRmtSvrEvt to handle incoming eventRmtSvrOnline and eventRmtSvrOffline.

dbusprxy.cpp: made changes to OnRmtSvrOnOffline to pass the dbus event message
for `eventRmtSvrOffline/eventRmtSvrOnline' to the upper port.

iftasks.cpp: added check in CIfTaskGroup::OnChildComplete to find the pointer to
CIoManager via propIoMgr, so that, CIfTaskGroup can be used beyond the objects
inherited from CRpcInterfaceBase, such as port objects.

iftasks.cpp, reqfwdr.cpp, iftasks.h: added FindTaskByClsid to
CIfParallelTaskGrp, so that the CRpcReqForwarderProxy can cancel the io tasks on
the eventRmtSvrOffline from the CRpcTcpBridge.