ifhelper.h, iftasks.cpp, iftasks.h: added CIfDeferCallTask and DEFER_IFCALL_NOSCHED to wrap the CDeferredCall to a task that can run within a taskgroup, so that the deferred calls can be run sequencially via a CIfTaskGroup. Accordingly the CIfTaskGroup is also modified for this purpose, that is,     . A task state is added to the CIfTaskGroup with three three states,     stateStarting, stateStarted, stateStopped.
    . The propRunning is now removed when the CIfTaskGroup cannot accept new
    tasks. the removal of propRunning guarantees no new tasks can be added any
    more.

    . AppendTask is added more check to prevent new tasks if the taskgroup is
    either stateStopped or not IsRunning but still stateStarted.

    . now propRunning, propNoResched, propCanceling and the task state are
    combined to be the taskgroup's state machine.

    . CIfParallelTaskGrp is also changed accordingly for the newly added task
    state.

rpcif.h, proxy.h, rpcroute.cpp, rpcroute.h :
    . Added AddSeqTask and m_pSeqTasks to support sequential operations. The
    sequential tasks can be added to the taskgroup of m_pSeqTasks, which is
    guarantted to service the tasks in the first come first service.

    . CRpcRouter has its OnRmtSvrOffline and OnRmtSvrOnlie sequentially executed
    with AddSeqTask. The reason is sometimes the eventRmtSvrOffline can come
    immediately after eventRmtSvrOnline and some of the registrations by
    OnRmtSvrOnline are not ready yet, and cause the OnRmtSvrOffline failed to do
    the removal and the CRpcTcpBridgeImpl and the CTcpStreamPdo leaks.

autoptr.cpp: added serialization logic for all the CObjBase's subclass, although
the implementation depends on each individual subclass. is it a security hole?

clsids.h, tcpport.h, tcpfido.cpp: added CTcpFidoListenTask to CRpcTcpFido to
listen on the underlying stream pdo, and receive the incoming requests, the same
as CDBusProxyFdo does.  the CRpcTcpFido is now an FDO instead.

port.cpp, port.h, tcpfido.cpp: added FindPortByType to be able to find the pdo
port on the port stack for firing eventRmtSvrOffline event when all the ports
are ready on the stack.

bridge.cpp: fixed a bug that SetupReqIrp did not setup the irp for
IF_METHOD_DISABLEEVT and IF_METHOD_ENABLEEVT are missing, because
CRpcTcpBridgeProxy's SetupReqIrp missed to call super::SetupReqIrp

iomgr.cpp: recovered the removal of unloadable ports in CIoManager::ClosePort by
an additional check whehter the iomanager is quitting. the port removal cannot
be called twice. otherwise, unexpected.

irp.h, irp.cpp, portstat.cpp: initialized m_dwCurPos to zero, to avoid
uninitialized m_dwCurPos referenced by portstat.cpp. Though the modification in
portstat also removed the dependency of m_dwCurPos by changing from
pIrp->MajorCmd() to pIrp->GetTopStack()->GetMajorCmd();

pnpmgr.cpp: made a change to allow the eventRmtSvrOnline to be broadcasted from
the pnpmgr. Previously it was considered to be no use. But now eventRmtSvrOnline
becomes important as the CRpcRouter needs the event to create the
CRpcTcpBridgeImpl.

defines.h: correct an error in the usage of sem_getvalue, but no effect since
Linux does not support negative semaphore value for #blocked processes.

dbusprxy.cpp: fixed a bug in HandleRmtRegMatch by wrapping the match object in a
CConfigDb object so that the CRpcReqForwarderImpl can handle the message via the
function map.

dbusprxy.cpp: fixed a bug in completion of SYS_EVENT_RMTSVREVENT, which removed
the incoming messages and blocked the upper port to receive the signal message.