stmpdo.cpp: added missing handler for CTRLCODE_RMTSVR_OFFLINE_PDO and CTRLCODE_RMTMOD_OFFLINE_PDO
iftasks.cpp: fixed a bug in CIfStartRecvMsgTask::StartNewRecv, which added too
many listening irps to the dbuslocal port when StartRecvMsg returns pending.

iftasks.cpp, ifhelper.h: added CIfDeferredHandler, DEFER_HANDLER_NOSCHED, and
NewDeferredHandler. The CIfDeferredHandler can service as a function wrapper
with the Taskgroup support, so that the interface handler can be sequentially
executed by wrapping the handler with the CIfDeferredHandler and sending to the
AddSeqTask. The CRpcReqForwarder's OpenRemotePort and CloseRemotePort now use this
technique for sequential execution. the difference from the CIfDeferCallTask is
that, the CIfDeferredHandler accepts the callback from the handler function,
serves as a broker to copy the propMsgPtr and propReqPtr for handler's access,
and copy the returned propRespPtr for the callback access.

pnpmgr.cpp: made the change that QueryStop's return code is always ignored
except STATUS_PENDING, to make sure the DoStop always has chance to be executed.

rpcpdo.cpp, prxyfdo.cpp: fixed two segment fault bugs.

rpcroute.cpp: sequentialized the OnRmtSvrOnline function call instead of the
prevous CRouterOpenRmtPortTask.

pnpmgr.cpp: added an CIfDummyTask as the propEventSink as is needed when the
CPnpMgrStartPortCompletionTask tries to remove the CTcpStreamPdo on start
failure in StartPortsInternal. Without this eventSink ptr, the
CPnpMgrStartPortCompletionTask cannot be started and the port cleanup procedure
won't happen and the `socket-closed' CTcpStreamPdo will leak. Till now, i think
we have fixed the CRpcTcpBridgeImpl leaking issue. To sum up, this bug has three
issue combined,

    . The OpenRemotePort and CloseRemotePort need to be executed sequentialized
    at least on the same dbus session.

    . The eventRmtSvrOffline can only be sent out if the eventRmtSvrOnline has
    be sent out. otherwise, the eventRmtSvrOffline could come ahead of the
    eventRmtSvrOnline, and cause the CRpcTcpBridgeImpl objects to leak.

    . The missing eventSink ptr caused CPnpMgrStartPortCompletionTask failed to
    run to clean up the stopped CTcpStreamPdo.