. iftasks.cpp, iftasks.h: Changed the m_pParentTask in CIfRetryTask from a pointer to an CAutoPtr, to prevent the taskgroup to go away before the child task completes.
. iftasks.cpp, rpcif.cpp: added MarkPending in CIfTaskGroup's code for the tasks
unable to run immediately. There were times the tasks are scheduled to the other
thread and the Pending flag was not set, and caused the callback for the task
cannot be called and result in endless wait. As happened in CIoReqSyncCallback
on calling proxy method `Method_Echo'.

. irpthread.cpp: fixed a bug that CoGetClassName could return nullptr and caused
the std::string's constructor to segment fault.

. rpcroute.cpp: fixed a bug by adding the missing callback to the
CRouterOpenRmtPortTask, which could cause the sequential taskgroup to stop
flowing, and tasks piling in the group.

. stmpdo.cpp: improved checking of the socket state in OnPortStackReady, to make
sure the port get removed if the socket was already stopped.

. reqfwdr.cpp: made the EnableDisableEvent to run sequentially instead of
concurrently, so that the match registration/unregistration has a definite
result.

. tcpport.cpp: made the change to enable nonblocking transfer for the passive
opened socket in CRpcStreamSock's Start method

. tcpport.cpp: made the change in FillPacket to remove the packet header from
the payload buffer. and Also fixed the error in the COutgoingPacket::StartSend
to stop sending the packet header twice.

. tcpport.cpp: Fixed the error in CRpcStream::HandleReadIrp, that the packets
in the queue are not removed after dispatched to the client, and causes the
CTcpFidoListenTask to endlessly receive the same data.

. tcpport.cpp: added two missing field in CPacketHeader's initialization in
COutgoingPacket::SetHeader.

. irpthread.cpp: fixed a deadlock issue, when the RescheduleTask is running
a taskthread, waiting for the threadpool lock, while the threadpool's has
acquired the threadpool lock and waiting for the taskthread to stop.

. tcpfido.cpp: reviewed and fixed some bugs in CheckAndSend of CTcpStreamPdo,
and StartSend's of the CRpcStreamSock, COutgoingPacket and CRpcStream.

. reqfwdr.cpp: fixed a bug in CReqFwdrEnableRmtEventTask's OnTaskComplete, to
check the response parameter pack rather than the input parameter `iRetVal' to
determine the return code from the remoter server.

. tractgrp.h, rpcroute.cpp: fixed an object leak bug caused by the wrong
AppendTask is called in the CIfTransactGroup in the constructor.