. rpcroute.cpp: Made the bridge-side EnableRemoteEvent to run sequentially and fixed some bugs in the group of tasks for bridge-side event relay setup.
. tractgrp.h: Fixed wrong error code produced by CIfTransactGroup's
OnTaskComplete, which affected the error passed back to the client.

. iomgr.cpp, iftasks.cpp, ifstat.cpp, ifstat.h, iftasks.cpp, proxy.h: made
changes to the ClosePort of CIoManager and CRpcBaseOperations and
CInterfaceState by adding a callback parameter, so that the port stop and
destroy completion can be notified to the caller directly instead of polling
repeatedly. The callback is hooked via the newly added CloseChildPort from
CIoManager, which calls IRP_MN_PNP_STOP_CHILD, to stop the specified port by its
parent bus port.  And moved CBusPortStopSingleChildTask from port.cpp to
iftasks.cpp because it inherits from CIfRetryTask.

. port.cpp, stmpdo.cpp, rpcpdo.cpp, tcpfido.cpp, dbusprxy.cpp, dbusport.cpp,
stmpdo.cpp, tcpport.cpp: Reviewed the code and changed some occurances of
IrpCtxPtr& to IrpCtxPtr to avoid unexpected behavior when the irp's context
stack is changed.

.dbusport.cpp, evloop.cpp, evloop.h, mainloop.h: made change to the
CDBusBusPort::Stop to detach the dbus connection from the mainloop. Accordingly,
StopDBusConn is added to CEvLoop

. rpcpdo.cpp: fixed a concurrent bug in DispatchRespMsg and DispatchReqMsg, when
the request message arrives sooner before the SubmitIrp call quit, and the irp
status was overwritten to STATUS_PENDING by the submitIrp after the irp has been
set to complete, and caused the irp unable to finish the completion and the
whole incoming flow stalled in CRpcReqForwarderImpl.

. rpcif.cpp: Rewrote the CRpcInterfaceBase::AddAndRun to fix a concurrent issue
that the root task group and parallel task group could possibly stall the task
flow in some rare condition.

. iomgr.cpp, irpthread.cpp: Added code to fix two memory leaks. One is to remove
the code, which clear the irp's port pointer on the top context stack, which
will be used to release the port specific m_pExtBuf from the context. the other
is to delete the m_pServiceThread of the COneshotTaskThread on destruction.

. reqfwdr.cpp: fixed a bug in EnableDisableEvent where the
CReqFwdrEnableRmtEventTask is marked pending before it runs, which can cause
crash if the task returns immediately.

. stmpdo.cpp: added a 10 ms sleep for the OnNewConnection EAGAIN failure with
PORT_STATE_BUSY_SHARED. This could have preformance impact if it happens too
frequently.

. ifhelper.h: added DEFER_CALL_DELAY for a scheduled deferred call.

. reqfwdr.cpp: in RunTaskInternal of CReqFwdrOpenRmtPortTask, added code to stop
the interface on Start failure, to release the resources.

. iftest.cpp: Made changes to let the interface be released on Echo error,
CPPUNIT_ASSERT will skip the cleanup.