. rpcif.cpp, iftasks.cpp: Removed DelayRun in OnCancel of CIfTaskGroup and CIfParallelTaskGrp, which could led to memory leak due to the timerservice is no longer running. And as a alternative solution, the OnCancel routine honors STATUS_PENDING and let the caller to decides what to do. And in ClearActiveTasks, the approach on get status pending is to wait till the CIfRootTaskGroup returns other error code than STATUS_PENDING.
. buffer.cpp, buffer.h: Added a 64 bytes internal buffer for CBuffer object, so
that the data requiring less than 64 bytes can be held in the internal buffer to
avoid an extra malloc and memory fragmentation.

. reqfwdr.cpp, rpcroute.cpp: Changed the AddSeqTask to run on the CTaskThread
instead of the oneshot thread, to give better memory performance, that is,
stable memory volume and lower rising speed of memory.

. tcpport.cpp: refacted the CRpcStream's GetReadIrpsToComp with better locking
solution, and better stream support, that is, better support for reading
requested size of bytes from the stream beyond the packet boundary.

. rpcif.cpp: Removed the lock of the task to run when inspecting the error code
of the task in AddAndRun, which could result in deadlock due to the opposite
locking sequence to that of task completion.

. rpcif.cpp, bridge.cpp, proxy.h, stream.cpp: changed ToPublicName to
ToPublicName_NoStr in hope it could reduce the memory leak. and also replaced
the std::string operation with sprintf and local buffer for GetIidFromIfName for
the same purpose. the effect is not significant.

. dmsgptr.cpp: added some debugging code for dbus message dumping to local file.
. objbase.cpp: changed to DumpObj's output formant.

. ipc/Makefile.in,  portex.cpp: added portex.cpp to the project.

. dbusport.cpp, dbusport.h: added AsyncStopDBusConn and PreStop to the
CDBusBusPort to remove the dbus connection gracefully on the main loop thread
instead of the task thread previously, which could result in segment fault due
to lack of synchroniztion, or unexpected halt for 30 secs.

. ifstat.cpp: Unsubscribed the events before close the port to prevent feedback
event during port closing.

. iftasks.cpp, iftasks.h: added OnChildComplete for CIfRootTaskGroup to prevent
rescheduling of the root task  when the task queue is empty, which could cause
rerun of the task if the new task was added immediately afterward.

. iftasks.cpp: in CIfRetryTask::Process, if eventCancelTask finds the OnCancel
returns STATUS_PENDING, it will not overwrite it with ERROR_CANCEL, which is
especially true for CIfTaskGroup and CIfParallelTaskGrp if two threads running
OnCancel on the same task group at the same time.

. irp.cpp: the irp is now initialized with its status set to STATUS_PENDING.

. port.cpp: refacted CGenBusPortStopChildTask to have better processing when the
port statck is not stopped sucessufully with an -EAGAIN return code. When it
happens, the task will be rescheduled to retry in 1 seconds to stop the port
stack again. it is not very effective if the failure happens at the non-top
ports or the port's state is not PORT_STATE_READY any more.

. port.cpp, portstat.h, portstat.cpp, portex.cpp: added support to
PORT_STATE_STOP_RESUME and PORT_STATE_BUSY_RESUME which can help to resume a
stop later if the port current is in a temporary busy state. The example is
that, the dbuslocalpdo is handling a incoming message in the
PORT_STATE_BUSY_SHARED state, while the stop arrives. And the stop will fail in
the CPort::CanContinue, which will put the port in the
PORT_STATE_EXCLUDE_PENDING states to wait till all the outstanding
PORT_STATE_BUSY_SHARED requests to complete and then resume the stop process.
To support this, CPortState is added HandleBusyResume and HandleStopResume, and
the SetPortState, PopState and CanContinue are changed to support this function.
And the resume function pointer is replaced with a taskletptr which have more
flexibility.

. iftest.cpp: removed some of CPPUNIT_ASSERT which could skip the pIf's Stop
process and causes unexpected result.