. buffer.cpp, buffer.h: fixed a bug that sometimes the buffer assignment missed to update the `data type' field, which caused the a `byte array' buffer to be released as a `DMsgPtr' buffer.
. ifstat.cpp: refacted the CInterfaceState::ClosePort

. iftasks.cpp: made change to the CIfStopTask to switch interface state via the
OnPortEvent from the interface object, instead of SetStateOnEvent, which does
more cleanup work.

. ifstat.cpp, reqfwdr.cpp, rpcroute.cpp, iftasks.cpp, bridge.cpp, rpcif.cpp:
Introduced `stateStartFailed' state to the interface state machine, which will
label the interface state when starting interface failed. It is added to tell
the afterwards stopping that the interface is to be stopped due to a failed
start, unlike the successful start. And all the call to
CRpcInterfaceBase::StopEx needs to first SetStateOnEvent before it can move on,
to prevent StopEx from being called twice.

. drvmgr.cpp: fixed a memory leak that the Json::CharReader is not deleted after
it is done.

. ifhelper.h, proxy.h, rpcif.cpp, bridge.cpp, stream.cpp: fixed a memory leak by
removing ToPublicName and ToInternalName and the related conversions for the
CoGetIidFromIfName or CoGetIfNameFromIid. The instance-wise interface internal
name has produced too many useless entries in the g_oIidDict beyond the
lifecycle of the interface object.

. iomgr.cpp: replaced the iteration of the irp's context stack, in CIoManger's
CompleteIrp and CancelIrp, from the iterator to the integer index to avoid the
unexpected reference error if the irp's context stack( std vector ) has its
entry poped in the CPort::CompleteIrp or CPort::CancelIrp.

. iomgr.cpp, ptifmap.cpp: Guarded HandleToPort/PortToHandle with the
CIoManager::GetPortPtr because the port object could have been removed and the
HandleToPort could generate an pointer to an invalid memory area. With the
GetPortPtr, the port is guarantted to be valid and won't be removed during the
call.

. pnpmgr.cpp: CPnpMgrStopPortAndDestroyTask is now scheduled to run on task
thread but not the oneshot thread, to improve the performance. the previous
concern about the synchronous call to StopPortStack has gone because
DestroyPortStack will always provide a marster irp, and so no long wait any
more.

. portstat.cpp: added code to enable the port to be destroyed at the attached
state, if it cannot start successfully.

. ptifmap.cpp: fixed a memory leak that the elements in m_setHandles have a
one-to-many mapping from port to the interface. So the element in m_setHandles
should be removed when the last interface associated with this port is deleted.
if not, those interface could deposit in the m_mapIf2Port and hanging there
    without being removed.

. rpcif.cpp: made improvement to access m_pRootTaskGroup safer by referencing
the group via an autoptr before member access.

. reqfwdr.cpp: fixed a bug in CReqFwdrOpenRmtPortTask::RunTaskInternal, which
called StopEx synchronously and made the CReqFwdrOpenRmtPortTask be reentered
during completion. The inline comment has detail explanation.

. stmpdo.cpp, tcpport.h: added a semaphore to CTcpStreamPdo, which is to make
sure the eventRmtSvrOnline strictly coming ahead of the eventRmtSvrOffline.
otherwise, weird things can happen.

. tcpport.cpp, tcpport.h: made changes to return -ENETDOWN rather than
ERROR_STATE if the socket is stopped.

. iftest.cpp: added a memory testing function. but it seems useless.