. objbase.cpp: enhanced the random number generation.
. defines.h, evloop.cpp, evloop.h, mainloop.h: changed the type of `HANDLE' from
guint32 to intptr_t, for better compatibility.

. dbushook.cpp: fixed a potential bug in CDBusTimerCallback's ToggleWatch, which
mistakenly used srcAsync to start/stop the timer source, resulted in no effect.
This fix is an attempt to fix a segment fault when the timer source is being
removed, that the pointer to the timer source is not aligned to the dword
boundary.

. rpcif.cpp: fixed a bug that the propRespPtr cannot be found to FillRespData,
after immediate return from submitIrp in the SendMethodCall, due to the removal
of the callback ptr from the IRP.

. rpcpdo.cpp: fixed a typo bug.

. sevpoll.cpp: fixed a potential bug.

. dbusprxy.cpp, tasklets.h, prxyport.h: added `reconnect' support to the
CProxyPdoConnectTask, to start reconnection if the remote server goes offline.
the Task is started when the CDBusProxyFdo's listening task requests listening
to the remote event. The pdo is added m_pConnTask, Reconnect and ClearConnTask
for this change. Also the CProxyPdoConnectTask is changed to inherit from the
CThreadSafeTask instead of the original CTaskletRetriable, because it needs
to handle cancel as well as irp completion.

. bridge.cpp: fixed 3 bugs
    a)  added the missing resp message if the CRpcTcpBridgeProxy::ForwardRequest
    encounters a successful immediate return.

    b)  Changed CRpcTcpBridgeProxy's FillRespDataFwrdReq to return immediately if
    the response message returns error rather than proceeds to retrieve more
    response data.

    c)  In CRpcInterfaceServer's DoInvoke, removed a race condition when the
    ForwardRequest returns pending and attempt to retrieving the req message,
    the CReqFwdrForwardRequestTask on then other thread could also trying to
    modify the req message for sending, which causes the request fail
    prematurely because dbus_message_iter_init returns false to indicate the
    req message has no arguments.

. reqfwdr.cpp: Fixed a bug in CReqFwdrForwardRequestTask::OnTaskComplete, by
filling the return value to the response to the client.