. dmsgptr.cpp: added check of the `dwSize' of SERI_HEADER_BASE to avoid crash on toxic packets.
. propids.h, iftasks.cpp: added propKAParamList for calls to the task on the
`eventKeepAlive'. It is added to prevent the normal propParamList is overritten
when a KA event arrives at the same other task events is in process.

. iftasks.cpp: differred the call to CIfInvokeMethodTask's OnKeepAliveRelay to
break the potential long path of callback chain from the proxy's CIfIoReqTask
via many CIfInterceptTasks to the CIfInterceptTasks. The nested task locks could
cause dead lock unexpectedly.

. iftasks.cpp: added a default OnKeepAlive handler to the CIfInterceptTasks to
pass on `eventKeepAlive'.

. iftasks.cpp: fixed a bug in CIfIoReqTask that the remote error was not passed
on to the client, and returned 0 instead.

. rpcif.cpp: added the dmsg's serial number as the second clue on the server
side for the client to locate the task to keep-alive. For IPC, the taskid is
used to find the task on the client side( proxy ). And for the
CRpcReqForwarderProxy, the message serial number is used for this purpose,
because the ReqForwarderProxy does not aware of the format of the payload in the
dmesg, so the serial number is the only clue for it to find the CIfIoReqTask to
KA.

. prxyfdo.cpp: removed some bad code which prevent the KA event to be dispatched
to the related interface objects.

. reqfwdr.cpp, rpcroute.h: refacted CRpcReqForwarderProxy's OnKeepAliveOrig and
added BuildKeepAliveMessage to manualy build the KA event for the home stretch.
The old code does not follow the later changed layout of the dmsg's.

. rpcroute.cpp: fixed a minor bug, which mistakenly failed the CheckEvtToFwrd
call even on success.

. actctest.cpp: CancelRequest is found to possibly arrive earilier than the
request to cancel and then fail if the bridge or reqfwdr are very busy. So
adjusted the code a little. It is not a fix, just to reduce the possibility.

. prsvr.cpp: fixed an apparent bug.