. iftasks.cpp: added eventIrpComp handler for CIfDummyTask, to clear the cached irp pointer on irp completion, otherwise, memory leaks can happen.
. rpcif.cpp: m_pSeqTasks can sometimes contains tasks not started after the
interface has already stopped. So added a cancel call on the m_pSeqTasks in the
CRpcServices::OnPostStop.

. stream.cpp: brough forward the RemoveUxStream from OnTaskComplete to the
RunTask of CIfStopUxSockStmTask, before the StopEx is called, to prevent further
usage of the CUnixSockStream object in the stopping process.

. uxport.cpp: fixed some typo bugs which caused crashs or hangs.

. uxstream.cpp: refacted task CIfUxSockTransTask, and make the reading task to
read once instead of continuous read to exhause all the incoming packets,
otherwise the incoming data can occupy the CPU for a long time and
starve the other tasks.

. uxstream.h: replaced PostUxStreamEvent with SendUxStreamEvent to avoid the
event be queued again at the far end of the m_pSeqTask when the load is high,
and also removed DEFER_CALL related task with direct calls.