. configdb.h: fixed a bug in CopyParams from CParamList
. buffer.h:  added processing for the missing typeByteArr
    CBuffer::SerializePrimeType,

. iftasks.cpp: fixed two bugs for non-dmsg request processing, which is for the
support of the CRpcControlStream requests over the tcp connection.

. bridge.cpp, rpcroute.h: made OnPostStart to work for CRpcTcpBridge and
CRpcTcpBridgeProxy.

. stmpdo.cpp, tcpport.cpp:
    1. refacted the CheckAndSend method.
    2. fixed some bugs to make the CRpcControlStream work over the connection.

. stmrelay.h: made the InitUserFuncs to be seen by the master InitUserFuncs and
get called, so that the IStream interface can be registered and let the
SendFetch_Proxy to be able to setup the interface information for
SENDDATA/FETCHDATA requests.

. routmain.cpp: brought forward the SetCmdLineOpt, to make sure the command line
options guaranteed to exist before the iomanger is started. Otherwise, the
CRpcTcpBusPort may not be able to create the listening socket due to missing the
information about the router role.