.streamex.cpp, streamex.h, stmcli.cpp, stmsvr.cpp: Fixed some bugs, mainly the flow control issues, bool parameter issue and garbage in the event buffer passed by POST_LOOP_EVENT. Now the CStreamSyncBase class support the client to interact whether asynchronously or synchronously. I must confess that the asynchronous API is not developer friendly. There are still some boundary issues to check.
. rpcif.cpp: removed `IsConnected' check for `RunManagedTask' and `AddAndRun'
for the consideration that CIfParallelTaskGrp is now running far more types of
    tasks than IO tasks as it was, and therefore some tasks can run even the
    interface object is not at the state `stateConnected'.

. mainloop.h: added code to clear the m_dwTid after the loop is done. And added
the destructor for CMainIoLoopT.

. buffer.h: made the exception message more precise.

. iftasks.cpp, ifhelper.h, stmrelay.cpp, streamex.cpp: added `OnCancel' to
CIfResponseHandler to free the resources it has claimed, when the major task it
hooked failed immediately.

. stream.cpp, uxstream.cpp: made the CIfUxListeningTask the only task to read
all the packets incoming, as we have done to the CStreamXXXRelay classes. The
reason is that, we cannot fully control the stmevtLoopStart to be the first
event if we have CIfUxListeningTask and CIfUxSockTransTask both to deliver event
packets and data packets respectively.