. sftest.cpp, sfsvr.cpp, sfcli.cpp, sftest.cpp, sfsvr.h, driver.json: rewritten the sftest from the SendFetchServer to the SimpFileServer, and replaced the base class from CFileTransferServer/CFileTransferProxy to CStreamServerSync and CStreamProxySync. The new solution uses stream interface to do the transfer, and give more control to the server/proxy, as well as the complexity :D. If you want to do file transfer, better uses this project as the start point.
. evloop.cpp, mainloop.h: added OnPreLoop and OnPostLoop before loop starts and
after loop quits. Some tasks are added to the mainloop's task queue before the
mainloop starts, and OnPreLoop makes sure these tasks to run after the mainloop
starts. And OnPostLoop will cleanup the remaining tasks in the task queue when
the mainloop quits.

. mainloop.h: added exit-loop task queue, which will be run immediately after
the mainloop quits.

. streamex.h: added a parameter `pStartTask' to StartLoop, so that the
pStartTask can run when the loop is started. This is a workaround to pass some
parameters to the objects running on the loop.

. buffer.cpp: added the deserialization for the missing type `typeByteArr'.
. config.cpp: added the check of the deserialization of the map elements.

. evloop.cpp, evloop.h, streamex.h, sevpoll.cpp: added the return value for the
CMainIoLoop, and accordingly, a return code is appended to the event `aevtStop'

. ifhelper.h: added SupportIid to CAggInterfaceXXX, to allow an interface to be
inherited by another interface, in current case, IStream is inherited by
CMyFileServer interface. And accordingly, the DEFINE_UNIQUE_PROXY_IMPL and
DEFINE_UNIQUE_HANDLER_IMPL are updated to call this method as well as GetIid to
determine if the interface of the given iid is supported by the give class.

. iomgr.cpp, pnpmgr.cpp, port.cpp: removed the check `RunningOnMainThread' of
preventing open/start/destroy Port from mainloop thread, as long as it does not
take long to wait on the mainloop thread. The restriction became buggy when the
WORK_ITEM thread are removed and the timer callback can only happen on the
mainloop, especially when CIfOpenPortTask retries OpenPort, the check will fail
all the retries.

. rpcif.cpp: Improved the code to return error response when FetchData fails
somewhere. In some situation, say the deserialization failed, the FetchData
returned silently without response to the client.

. stmrelay.h, stmrelay.cpp: Fixed a bug in OnOpenStreamComplete and
OnFetchDataComplete, that when error occurs, the response was not set to the
correct place and caused the Invoke task cannot return the error response
immediately and let the client quit on timeout.

. irp.h, streamex.cpp, iomgr.cpp: added IRP_SUBMITED and SetSubmited,
ClearSubmited, and IsSubmited to the IoRequestPacket class. The purpose is to
let irp has a state to distiguish whether the irp is in processing, so that it
is convenient for CIfStmReadWriteTask's RunTask to tell if it start the
processing of the irp at the request queue head, or wait for the irp already in
processing to complete. Previously CUnixSockStream has a bug in WriteStream
which failed to return STATUS_PENDING, and cloaked this issue, till WriteStream
was able to return PENDING.

. uxport.cpp: fixed a bug in SubmitWriteIrp, which missed to start watching on
the socket for writing complete event, and stall the write requests afterwards.

. sfcli.cpp: sfsvr.cpp, sfsvr.h, stmcli.cpp, stmsvr.cpp, stmsvr.h: added a
return code to the event stmevtRecvData, so that OnRecvData_Loop can tell if the
last data receiving request failed or not.